Design chrome extension 크기에 맞게 수정

+++ src/components/TodoItem/Cancel.tsx
@@ -5,19 +5,22 @@ const Cancel = ({ id }: { id: number }) => {
   const { dispatch } = useDispatchTodoIds()
 
   const handleDelete = () => {
-    dispatch({ type: "DELETE", targetId: id })
+    // popup overlay example
+    if (confirm("정말 지울거냐구!")) {
+      dispatch({ type: "DELETE", targetId: id })
+    }
   }
 
   return (
     <div
       onClick={handleDelete}
-      className="group/cancel invisible absolute right-2 top-1/2 -translate-y-1/2 rounded-full p-0.5 transition group-hover/item:visible hover:bg-slate-100/50"
+      className="group/cancel invisible absolute right-2 top-1/2 -translate-y-1/2 rounded-full p-0.5 transition group-hover/item:visible hover:bg-slate-100/30"
     >
       <svg
         xmlns="http://www.w3.org/2000/svg"
         viewBox="0 0 20 20"
         fill="currentColor"
-        className="h-8 w-8 fill-slate-600 hover:fill-slate-900"
+        className="h-4 w-4 fill-slate-600 hover:fill-slate-900"
       >
         <path d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z" />
       </svg>
@@ -25,5 +28,4 @@ const Cancel = ({ id }: { id: number }) => {
   )
 }
 
-export default Cancel
 export default memo(Cancel)
